# perf(events): store AutoCmds in a kvec_t