Replace linked lists with kvec_t.

A quick benchmark on 10k autocommands of the same type and pattern (I doubt there is any performance gain with any other scenario, for now):
```
master:
nvim_create_autocmd     103.056ms
nvim_exec_autocmds      4.408ms
nvim_del_autocmd        273.789ms

refactor:
nvim_create_autocmd     10.079ms
nvim_exec_autocmds      4.420ms
nvim_del_autocmd        81.586ms
```

<details>

```lua
local api = vim.api
local hrtime = vim.loop.hrtime

require('jit').off()
require('jit').flush()

local N = 10000
local ts

local ids = {}
for i = 1, N do
  ids[i] = 0
end

local nvim_create_autocmd = api.nvim_create_autocmd
ts = hrtime()
for i = 1, N do
  ids[i] = nvim_create_autocmd('User', {
    pattern = 'Benchmark',
    command = 'eval ' .. i,
  })
end
print(('nvim_create_autocmd     %.3fms'):format((hrtime() - ts) / 1000000))

ts = hrtime()
api.nvim_exec_autocmds('User', { pattern = 'Benchmark', modeline = false })
print(('nvim_exec_autocmds      %.3fms'):format((hrtime() - ts) / 1000000))

local nvim_del_autocmd = api.nvim_del_autocmd
ts = hrtime()
for i = 1, N do
  nvim_del_autocmd(ids[i])
end
print(('nvim_del_autocmd        %.3fms'):format((hrtime() - ts) / 1000000))

vim.cmd('qa!')
```

`nvim --headless -c 'so benchmark.lua'`

</details>