@justinmk 2023-04-20
Great to see improvements to the autocmd system. I think the right direction for https://github.com/neovim/neovim/issues/19156 is to massage the autocmds subsystem so it becomes a mature, generalized interface for everything we need.